feat(cli): add 'rnx-dep-check' command
