+++ change/@rnx-kit-cli-c683d1ef-b4ea-4d9b-bb26-185c2186ca93.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Added `rnx-dep-check` command",
+  "packageName": "@rnx-kit/cli",
+  "email": "4123478+tido64@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
