+++ packages/cli/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@rnx-kit/config": "0.2.5",
+    "@rnx-kit/dep-check": "^1.0.2",
     "chalk": "^4.1.0"
   },
   "devDependencies": {
