+++ packages/cli/src/parsers.ts
@@ -1,0 +1,27 @@
+import { AllPlatforms } from "@rnx-kit/config";
+
+export function parseBoolean(val: string): boolean {
+  if (val === "false") {
+    return false;
+  }
+  if (val === "true") {
+    return true;
+  }
+  throw new Error(
+    "Invalid boolean value '" + val + "' -- must be true or false"
+  );
+}
+
+export function parsePlatform(val: string): AllPlatforms {
+  switch (val) {
+    case "android":
+    case "ios":
+    case "macos":
+    case "win32":
+    case "windows":
+      return val;
+
+    default:
+      throw new Error("Invalid platform '" + val + "'");
+  }
+}
